pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["alexeygalt <alexeygalt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gunicorn = "^21.2.0"
pydantic = "^2.3.0"
python-dotenv = "^1.0.0"
pre-commit = "^3.3.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
